qt: rebuild (simpler cpp-side array parameters)

+++ qt/gen_qabstractitemmodel.cpp
@@ -576,7 +576,7 @@ void QAbstractItemModel_DataChanged3(QAbstractItemModel* self, QModelIndex* topL
 	roles_QList.reserve(roles->len);
 	int* roles_arr = static_cast<int*>(roles->data);
 	for(size_t i = 0; i < roles->len; ++i) {
-		roles_QList.push_back(roles_arr[i]);
+		roles_QList.push_back(static_cast<int>(roles_arr[i]));
 	}
 	self->dataChanged(*topLeft, *bottomRight, roles_QList);
 }
