+++ qt/gen_qcolumnview.cpp
@@ -107,7 +107,7 @@ void QColumnView_SetColumnWidths(QColumnView* self, struct miqt_array* /* of int
 	list_QList.reserve(list->len);
 	int* list_arr = static_cast<int*>(list->data);
 	for(size_t i = 0; i < list->len; ++i) {
-		list_QList.push_back(list_arr[i]);
+		list_QList.push_back(static_cast<int>(list_arr[i]));
 	}
 	self->setColumnWidths(list_QList);
 }
