+++ qt/gen_qpen.cpp
@@ -73,7 +73,7 @@ void QPen_SetDashPattern(QPen* self, struct miqt_array* /* of double */ pattern)
 	pattern_QList.reserve(pattern->len);
 	double* pattern_arr = static_cast<double*>(pattern->data);
 	for(size_t i = 0; i < pattern->len; ++i) {
-		pattern_QList.push_back(pattern_arr[i]);
+		pattern_QList.push_back(static_cast<double>(pattern_arr[i]));
 	}
 	self->setDashPattern(pattern_QList);
 }
