+++ qt/gen_qtreeview.cpp
@@ -357,7 +357,7 @@ void QTreeView_DataChanged3(QTreeView* self, QModelIndex* topLeft, QModelIndex* 
 	roles_QList.reserve(roles->len);
 	int* roles_arr = static_cast<int*>(roles->data);
 	for(size_t i = 0; i < roles->len; ++i) {
-		roles_QList.push_back(roles_arr[i]);
+		roles_QList.push_back(static_cast<int>(roles_arr[i]));
 	}
 	self->dataChanged(*topLeft, *bottomRight, roles_QList);
 }
